+++ modelopt/torch/quantization/nn/modules/tensor_quantizer.py
@@ -747,8 +747,9 @@ class TensorQuantizer(nn.Module):
             reduce_axis = quant_utils.convert_quantization_axis_to_reduce_axis(inputs, self._axis)
             amax = quant_utils.reduce_amax(inputs, axis=reduce_axis, keepdims=True).detach()
 
-        amax = amax.detach() if is_torch_export_mode() else amax.data
-        return amax
+        if is_torch_export_mode() or torch.compiler.is_exporting():
+            return amax.detach()
+        return amax.data
 
     def validate_attr(
         self, attr_value=None, attr_name="amax", raise_error=False, warn_error=False, name=""